requirements.txt
absl-py==1.0.0
alembic==1.7.6
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bert-extractive-summarizer==0.10.1
black==21.12b0
bleach==4.1.0
blinker==1.4
blis==0.7.5
Brlapi==0.7.0
cachetools==4.0.0
catalogue==2.0.6
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.2
fasteners==0.14.1
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.4.0
google-api-python-client==1.7.11
google-auth==2.3.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
httplib2==0.14.0
huggingface-hub==0.4.0
idna==2.8
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.1
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==12.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
mpmath==1.2.1
murmurhash==1.0.6
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.6.7
nose==1.3.7
notebook==6.4.7
numpy==1.22.1
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.1
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.8.2
pydot==1.4.2
PyDrive==1.3.1
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
regex==2022.1.18
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.0
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.3.12
srsly==2.4.2
stack-data==0.1.4
sympy==1.9
systemd-python==234
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.11.4
tomli==1.2.3
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typer==0.4.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.1
wadllib==1.3.3
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
WTForms==3.0.1
xkit==0.0.0
zipp==3.7.0
